+++ CHANGELOG.md
@@ -18,5 +18,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for Parquet files
 - Added more entities and functionality
 
+## [0.1.2] - 2025-02-18
+
+### Added
+- Added new topic entity
+- Added base entity for CRUD testing
+- Added topic tests
+
 [0.1.0]: https://github.com/username/repository/releases/tag/v0.1.0
 [0.1.1]: https://github.com/username/repository/releases/tag/v0.1.1
+[0.1.2]: https://github.com/username/repository/releases/tag/v0.1.2